change structure of statistic

+++ src/main/java/com/travelstory/dto/statistic/LikeStatistic.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.query.Param;
 
 public interface LikeStatistic extends JpaRepository<Like, Long> {
     @Query(value = "SELECT count(id) from likes where MONTH (created_at) =:num", nativeQuery = true)
-    Long countLikeByCreatedAtMonth(@Param("num") Integer numberOfMonth);
+    Integer countLikeByCreatedAtMonth(@Param("num") Integer numberOfMonth);
 
 }
